fix: Redis service 구현체만 남김
- interface 불필요
