+++ wecare-backend/wecare/src/main/java/com/example/wecare/common/service/RedisService.java
@@ -1,9 +1,27 @@
 package com.example.wecare.common.service;
 
-import java.util.concurrent.TimeUnit;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Service;
 
-public interface RedisService {
-    void setValues(String key, String value, long timeout, TimeUnit unit);
-    String getValues(String key);
-    void deleteValues(String key);
+@Service
+@RequiredArgsConstructor
+public class RedisService {
+    private final RedisTemplate<String, String> redisTemplate;
+
+    public void setValues(String key, String value, long timeout) {
+        redisTemplate.opsForValue().set(key, value, timeout);
+    }
+
+    public String getValues(String key) {
+        return redisTemplate.opsForValue().get(key);
+    }
+
+    public void deleteValues(String key) {
+        redisTemplate.delete(key);
+    }
+
+    public boolean hasKey(String key) {
+        return redisTemplate.hasKey(key);
+    }
 }
