--- wecare-backend/wecare/src/main/java/com/example/wecare/common/service/RedisServiceImpl.java
@@ -1,30 +1,0 @@
-package com.example.wecare.common.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Service;
-
-import java.util.concurrent.TimeUnit;
-
-@Service
-@RequiredArgsConstructor
-public class RedisServiceImpl implements RedisService {
-
-    private final RedisTemplate<String, String> redisTemplate;
-
-    @Override
-    public void setValues(String key, String value, long timeout, TimeUnit unit) {
-        redisTemplate.opsForValue().set(key, value, timeout, unit);
-    }
-
-    @Override
-    public String getValues(String key) {
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    @Override
-    public void deleteValues(String key) {
-        redisTemplate.delete(key);
-    }
-
-}
